+++ RELEASE_NOTES.md
@@ -5,11 +5,17 @@ A new header is inserted each time a *tag* is created.
 
 ## main branch
 
+## v1.28.0
+
 - engine: LiSPSM is now a user settable option
+- engine: get the morph target buffer to the given primitive
 - Java: Fix TransformManager.getChildren()
 - Metal: newer devices are no longer limited to 16 samplers per Material.
 - gltfio: fix interpretation of occlusion strength
 - engine: minsdk is now 21 instead of 19. This allows the use of OpenGL ES 3.1
+- gltfio: fix ubershader issues with assignment of dummy textures
+- gltfio: material instances and variants are now accessed via `FilamentInstance` [⚠️ **API Change**]
+- gltfio: the animator can now only be accessed via `FilamentInstance` [⚠️ **API Change**]
 
 ## v1.27.2